Brute

Documentation for Brute.

Brute.Evaluation

Brute.@test

Brute.Evaluation — Type

Evaluation

Create an Evaluation object to store Tests. Madatory keywords are name and url.

Brute.@test — Macro

macro test(evaluation, stud_call, teach_call, settings)

Test stud_call vs teach_call to see if the outputs match. Results in a Test object. evalutation of type Evaluation has to be passed to save the test results in.

Additional Test setting passed as a named tuple to settings

Examples

@test evaluation student.f(1) student.f(2) (access=public)